Filter in usePersistedStateReducer
