+++ assets/src/hooks/usePersistedStateReducer.ts
@@ -17,10 +17,20 @@ const usePersistedStateReducer = (
   )
 
   useEffect(() => {
-    saveState(APP_STATE_KEY, state, PERSISTED_KEYS)
+    const perstableState = filter(state, PERSISTED_KEYS)
+    saveState(APP_STATE_KEY, perstableState)
   }, [state])
 
   return [state, dispatch]
+}
+
+export const filter = (obj: object, allowedKeys: string[]): object =>
+  keys(obj)
+    .filter(key => allowedKeys.includes(key))
+    .reduce((acc, key) => ({ ...acc, [key]: obj[key] }), {})
+
+function keys<O extends object>(obj: O): Array<keyof O> {
+  return Object.keys(obj) as Array<keyof O>
 }
 
 export default usePersistedStateReducer
