+++ assets/src/localStorage.ts
@@ -1,10 +1,5 @@
-export const saveState = (
-  stateKey: string,
-  state: object,
-  persistedKeys: string[]
-) => {
-  const perstableState = filter(state, persistedKeys)
-  const serializedState = JSON.stringify(perstableState)
+export const saveState = (stateKey: string, state: object) => {
+  const serializedState = JSON.stringify(state)
   localStorage.setItem(stateKey, serializedState)
 }
 
@@ -16,13 +11,4 @@ export const loadState = (stateKey: string): object | undefined => {
   }
 
   return JSON.parse(serializedState)
-}
-
-export const filter = (obj: object, allowedKeys: string[]): object =>
-  keys(obj)
-    .filter(key => allowedKeys.includes(key))
-    .reduce((acc, key) => ({ ...acc, [key]: obj[key] }), {})
-
-function keys<O extends object>(obj: O): Array<keyof O> {
-  return Object.keys(obj) as Array<keyof O>
 }
