+++ assets/tests/localStorage.test.ts
@@ -1,8 +1,6 @@
-import { filter, loadState, saveState } from "../src/localStorage"
+import { loadState, saveState } from "../src/localStorage"
 
 const APP_STATE_KEY = "test-mbta-skate-state"
-
-const PERSISTED_KEYS = ["selectedRouteIds"]
 
 const mockLocalStorage = {
   getItem: jest.fn(),
@@ -24,20 +22,7 @@ describe("saveState", () => {
 
   test("saves stringified state to local storage", () => {
     const state = { selectedRouteIds: ["28", "39"] }
-    saveState(APP_STATE_KEY, state, PERSISTED_KEYS)
-
-    expect(mockLocalStorage.setItem).toHaveBeenCalledWith(
-      APP_STATE_KEY,
-      '{"selectedRouteIds":["28","39"]}'
-    )
-  })
-
-  test("only saves defined saved keys", () => {
-    const state = {
-      foo: "bar",
-      selectedRouteIds: ["28", "39"],
-    }
-    saveState(APP_STATE_KEY, state, PERSISTED_KEYS)
+    saveState(APP_STATE_KEY, state)
 
     expect(mockLocalStorage.setItem).toHaveBeenCalledWith(
       APP_STATE_KEY,
@@ -77,34 +62,5 @@ describe("loadState", () => {
       .mockImplementation((_stateKey: string) => null)
 
     expect(loadState(APP_STATE_KEY)).toBeUndefined()
-  })
-})
-
-describe("filter", () => {
-  const originalLocalStorage = window.localStorage
-
-  beforeAll(() => {
-    Object.defineProperty(window, "localStorage", {
-      value: mockLocalStorage,
-    })
-  })
-
-  afterAll(() => {
-    Object.defineProperty(window, "localStorage", originalLocalStorage)
-  })
-
-  test("filters an object by allowed keys", () => {
-    const originalObject = {
-      one: 1,
-      two: 2,
-      three: 3,
-    }
-    const allowedKeys = ["one", "three"]
-    const expected = {
-      one: 1,
-      three: 3,
-    }
-
-    expect(filter(originalObject, allowedKeys)).toEqual(expected)
   })
 })
